Passing AddSong through component tree: completed, and now alert status can be modified and styled
