+++ frontend/song-on-spotify/src/App.js
@@ -4,23 +4,21 @@ import { Route, Switch } from 'react-router-dom';
 import Home from './components/home';
 import Navigation from './components/navigation';
 
-class App extends React.Component {
-    render() {
-        return (
-            <div className='App'>
-                <header>
-                    {/* Navigation */}
-                    <Navigation />
-                </header>
+const App = () => {
+    return (
+        <div className='App'>
+            <header>
+                {/* Navigation */}
+                <Navigation />
+            </header>
 
-                <main className='content'>
-                    <Switch>
-                        <Route exact path='/' component={Home} />
-                    </Switch>
-                </main>
-            </div>
-        );
-    }
-}
+            <main className='content'>
+                <Switch>
+                    <Route exact path='/' component={Home} />
+                </Switch>
+            </main>
+        </div>
+    );
+};
 
 export default App;
