+++ frontend/song-on-spotify/src/components/home.js
@@ -4,12 +4,16 @@ import axios from 'axios';
 import { useAccessToken } from '../hooks/useAccessToken';
 import Playlists from './playlists';
 import Search from './search';
+import Alert from './alert';
 
 const Home = () => {
     const [userData, setUserData] = useState(null);
     const [accessToken, setAccessToken] = useAccessToken(
         window.location.search
     );
+
+    const [status, setStatus] = useState(null);
+    const [closed, setClosed] = useState(false);
 
     useEffect(() => {
         if (!userData) {
@@ -36,9 +40,44 @@ const Home = () => {
             .catch((err) => console.log('ERROR GET /v1/me', err));
     };
 
+    const addSong = async (e, uri) => {
+        e.preventDefault();
+
+        let playlistId = localStorage.getItem('playlist_id');
+        console.log(playlistId, uri, accessToken);
+
+        if (accessToken) {
+            //  https:stackoverflow.com/questions/60811947/how-to-pass-authorization-token-in-header-to-react-axios-post
+            await axios
+                .post(
+                    `https:api.spotify.com/v1/playlists/${playlistId}/tracks?uris=${uri.toString()}`,
+                    null,
+                    {
+                        headers: {
+                            Authorization: 'Bearer ' + accessToken,
+                            'Content-Type': 'application/json',
+                        },
+                    }
+                )
+                .then((response) => {
+                    console.log('ADDING Song Response', response);
+                    setStatus('success');
+                })
+                .catch((err) => {
+                    console.log('ERROR POST /v1/playlists/:id/tracks', err);
+                    setStatus('fail');
+                });
+        } else {
+            console.log('no access token');
+        }
+    };
+
     if (userData) {
         return (
             <div className='bg-white px-6 my-6'>
+                {status !== null && closed !== true ? (
+                    <Alert status={status} />
+                ) : null}
                 <h1 className='font-bold text-2xl capitalize'>
                     Add a song to your playlist
                 </h1>
@@ -46,7 +85,7 @@ const Home = () => {
                     Search a song name and select what playlist <br /> you would
                     like to add it to!
                 </span>
-                <Search />
+                <Search addSong={addSong} />
             </div>
         );
     } else {
